Pokemon class no longer needs dungeon argument
There are now 2 subclasses of Pokemon, UserPokemon and EnemyPokemon.

This is because the way they are loaded is different.

The game now loads dungeon data and enemy data from xml rather than text files. Pokemon spawns are weighted.

User can now have multiple of the same type of pokemon, thanks to the user_id attribute instead of poke_id.