StartDungeonScene class
The flow of the program is as follows:
In MainMenuScene, when the user starts a new game, StartDungeonScene will begin.

StartDungeonScene will load the appropriate dungeon data, and then immediately begin FloorTransitionScene

FloorTransitionScene will generate the floor while presenting to the user the name of the dungeon and the floor they are entering. When the floor has finished generating, the DungeonScene will start

DungeonScene is where the gameplay occurs. On reaching the stairs, if there is another floor, the FloorTransitionScene will start again, and repeat the earlier steps. Otherwise, return to MainMenuScene